% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_triads.R
\name{calc_num_triads}
\alias{calc_num_triads}
\title{Calculates the number of triad}
\usage{
calc_num_triads(G)
}
\arguments{
\item{G}{igraph object}
}
\value{
list containing number of triads for the given igraph object
}
\description{
Calculates the number of triads for the given graph object
}
\examples{
triad_num <- calc_num_triads(signed_adjacency_matrix(1,50,0.5)$G)
triad_num02 <- calc_num_triads(G)
}
