% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeiderEnergy.R
\name{HeiderEnergy}
\alias{HeiderEnergy}
\title{Calculate the Heider energy}
\usage{
HeiderEnergy(Adj, triad_num, J = 1)
}
\arguments{
\item{Adj}{Signed adjacency matrix}

\item{triad_num}{number of triads for the graph object of the signed
adjacency matrix}

\item{J}{Heider interaction strength. The default value is 1}
}
\value{
Heider energy for the given signed adjacency matrix
}
\description{
Calculate the Heider energy
}
\examples{
E <- HeiderEnergy(Adj,triad_num)
}
