% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_adjacency_matrix.R
\name{signed_adjacency_matrix}
\alias{signed_adjacency_matrix}
\title{Generate a signed adjacency matrix}
\usage{
signed_adjacency_matrix(N, p, q = 1)
}
\arguments{
\item{N}{Number of nodes}

\item{p}{connection probability for Erdos Renyi graph. The default value is 1}

\item{q}{probability to form positive edge.}
}
\value{
A list of signed adjacency matrices and the corresponding graph
         object
}
\description{
Generate a signed adjacency matrix
}
\examples{
signed_adjacency_matrix(N=10, p=1)
signed_adjacency_matrix(N=50, p=0.8, q=0.5)
}
